# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Flowers Shop Documentation API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/users": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Все пользователи",
                "operationId": "2de50d9685640f415a9c02001843b92c",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Misha"
                                                    },
                                                    "surname": {
                                                        "type": "string",
                                                        "example": "Surname"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "email@mail.ru"
                                                    },
                                                    "age": {
                                                        "type": "integer",
                                                        "example": 18
                                                    },
                                                    "address": {
                                                        "type": "string",
                                                        "example": "city Taganrog"
                                                    },
                                                    "balance": {
                                                        "type": "integer",
                                                        "example": 525
                                                    },
                                                    "gender": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "isAdmin": {
                                                        "type": "boolean",
                                                        "example": true
                                                    },
                                                    "basket": {
                                                        "properties": {
                                                            "products": {
                                                                "type": "array",
                                                                "items": {
                                                                    "properties": {
                                                                        "id": {
                                                                            "type": "integer",
                                                                            "example": 1
                                                                        },
                                                                        "flower": {
                                                                            "properties": {
                                                                                "id": {
                                                                                    "type": "integer",
                                                                                    "example": 1
                                                                                },
                                                                                "title": {
                                                                                    "type": "string",
                                                                                    "example": "Title"
                                                                                },
                                                                                "description": {
                                                                                    "type": "string",
                                                                                    "example": "Description"
                                                                                },
                                                                                "count": {
                                                                                    "type": "integer",
                                                                                    "example": 10
                                                                                },
                                                                                "price": {
                                                                                    "type": "integer",
                                                                                    "example": 199
                                                                                },
                                                                                "category": {
                                                                                    "properties": {
                                                                                        "id": {
                                                                                            "type": "integer",
                                                                                            "example": 1
                                                                                        },
                                                                                        "title": {
                                                                                            "type": "string",
                                                                                            "example": "Title category"
                                                                                        }
                                                                                    },
                                                                                    "type": "object"
                                                                                },
                                                                                "tags": {
                                                                                    "type": "array",
                                                                                    "items": {
                                                                                        "properties": {
                                                                                            "id": {
                                                                                                "type": "integer",
                                                                                                "example": 1
                                                                                            },
                                                                                            "title": {
                                                                                                "type": "string",
                                                                                                "example": "Title tags"
                                                                                            }
                                                                                        },
                                                                                        "type": "object"
                                                                                    }
                                                                                },
                                                                                "images": {
                                                                                    "type": "array",
                                                                                    "items": {
                                                                                        "properties": {
                                                                                            "id": {
                                                                                                "type": "integer",
                                                                                                "example": 1
                                                                                            },
                                                                                            "title": {
                                                                                                "type": "string",
                                                                                                "example": "Title image"
                                                                                            }
                                                                                        },
                                                                                        "type": "object"
                                                                                    }
                                                                                }
                                                                            },
                                                                            "type": "object"
                                                                        },
                                                                        "count": {
                                                                            "type": "integer",
                                                                            "example": 1
                                                                        }
                                                                    },
                                                                    "type": "object"
                                                                }
                                                            },
                                                            "total_price": {
                                                                "type": "integer",
                                                                "example": 553
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/users/": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Создать",
                "operationId": "8d9199cf5da05b157f50ba6116510070",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "Name"
                                            },
                                            "surname": {
                                                "type": "string",
                                                "example": "Surname"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "email@mail.ru"
                                            },
                                            "password": {
                                                "type": "string",
                                                "example": "qwerty"
                                            },
                                            "age": {
                                                "type": "integer",
                                                "example": 53
                                            },
                                            "address": {
                                                "type": "string",
                                                "example": "address"
                                            },
                                            "gender": {
                                                "type": "integer",
                                                "example": 1
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}